Fix jsocket/jnetworkinterface

+++ jsocket/include/jinetaddress.h
@@ -79,7 +79,7 @@ class InetAddress : public virtual jcommon::Object{
 		 * \brief Get the address IPv4.
 		 *
 		 */
-		virtual std::vector<uint8_t> GetAddress();
+		virtual std::vector<uint32_t> GetAddress();
 
 };
