+++ jsocket/include/jinetaddress4.h
@@ -100,7 +100,7 @@ class InetAddress4 : public InetAddress{
 		 * \brief Get the address IPv4.
 		 *
 		 */
-		virtual std::vector<uint8_t> GetAddress();
+		virtual std::vector<uint32_t> GetAddress();
 
 };
